chore: Reorganize table components, types and util methods
