+++ src/components/Table/GridTable.tsx
@@ -1,89 +1,35 @@
 import memoizeOne from "memoize-one";
-import { observer } from "mobx-react";
-import React, { MutableRefObject, ReactElement, ReactNode, useContext, useEffect, useMemo, useRef } from "react";
-import { Link } from "react-router-dom";
+import React, { MutableRefObject, ReactElement, useEffect, useMemo, useRef } from "react";
 import { Components, Virtuoso, VirtuosoHandle } from "react-virtuoso";
-import { navLink } from "src/components/CssReset";
-import {
-  PresentationContextProps,
-  PresentationFieldProps,
-  PresentationProvider,
-} from "src/components/PresentationContext";
-import { useSetupColumnSizes } from "src/components/Table/columnSizes";
-import { createRowLookup, GridRowLookup } from "src/components/Table/GridRowLookup";
-import { GridSortContext, GridSortContextProps } from "src/components/Table/GridSortContext";
+import { PresentationFieldProps, PresentationProvider } from "src/components/PresentationContext";
 import { GridTableApi, GridTableApiImpl } from "src/components/Table/GridTableApi";
-import { RowStateContext } from "src/components/Table/RowState";
-import { SortHeader } from "src/components/Table/SortHeader";
-import { ensureClientSideSortValueIsSortable, sortRows } from "src/components/Table/sortRows";
-import { getTableStyles, GridStyleDef } from "src/components/Table/styles";
-import { SortOn, SortState, useSortState } from "src/components/Table/useSortState";
-import { Css, Margin, Only, Palette, Properties, Typography, Xss } from "src/Css";
+import { useSetupColumnSizes } from "src/components/Table/hooks/useSetupColumnSizes";
+import { SortState, useSortState } from "src/components/Table/hooks/useSortState";
+import {
+  Direction,
+  GridColumn,
+  GridDataRow,
+  GridTableXss,
+  Kinded,
+  ParentChildrenTuple,
+  RenderAs,
+  RowTuple,
+} from "src/components/Table/types";
+import { filterRows } from "src/components/Table/utils";
+import { createRowLookup, GridRowLookup } from "src/components/Table/utils/GridRowLookup";
+import { RowStateContext } from "src/components/Table/utils/RowState";
+import { sortRows } from "src/components/Table/utils/sortRows";
+import { defaultStyle, GridStyle, GridStyleDef, resolveStyles, RowStyles } from "src/components/Table/utils/styles";
+import { Css, Only } from "src/Css";
 import { useComputed } from "src/hooks";
 import { useRenderCount } from "src/hooks/useRenderCount";
-import { safeKeys } from "src/utils";
-import { getButtonOrLink } from "src/utils/getInteractiveElement";
-import { shallowEqual } from "src/utils/shallowEqual";
-import { defaultStyle } from ".";
-
-export type Kinded = { kind: string };
-export type GridTableXss = Xss<Margin>;
-
-export const ASC = "ASC" as const;
-export const DESC = "DESC" as const;
-export type Direction = "ASC" | "DESC";
-export const emptyCell: GridCellContent = { content: () => <></>, value: "" };
+import { Row } from "./components";
 
 let runningInJest = false;
 
 /** Tells GridTable we're running in Jest, which forces as=virtual to be as=div, to work in jsdom. */
 export function setRunningInJest() {
   runningInJest = true;
-}
-
-/** Completely static look & feel, i.e. nothing that is based on row kinds/content. */
-export interface GridStyle {
-  /** Applied to the base div element. */
-  rootCss?: Properties;
-  /** Applied with the owl operator between rows for rendering border lines. */
-  betweenRowsCss?: Properties;
-  /** Applied to the first non-header row, i.e. if you want to cancel out `betweenRowsCss`. */
-  firstNonHeaderRowCss?: Properties;
-  /** Applied to all cell divs (via a selector off the base div). */
-  cellCss?: Properties;
-  /**
-   * Applied to the header row divs.
-   *
-   * NOTE `as=virtual`: When using a virtual table with the goal of adding space
-   * between the header and the first row use `firstNonHeaderRowCss` with a
-   * margin-top instead. Using `headerCellCss` will not work since the header
-   * rows are wrapper with Chrome rows.
-   */
-  headerCellCss?: Properties;
-  /** Applied to 'kind: "totals"' cells */
-  totalsCellCss?: Properties;
-  /** Applied to the first cell of all rows, i.e. for table-wide padding or left-side borders. */
-  firstCellCss?: Properties;
-  /** Applied to the last cell of all rows, i.e. for table-wide padding or right-side borders. */
-  lastCellCss?: Properties;
-  /** Applied to a cell div when `indent: 1` is used. */
-  indentOneCss?: Properties;
-  /** Applied to a cell div when `indent: 2` is used. */
-  indentTwoCss?: Properties;
-  /** Applied if there is a fallback/overflow message showing. */
-  firstRowMessageCss?: Properties;
-  /** Applied on hover if a row has a rowLink/onClick set. */
-  rowHoverColor?: Palette;
-  /** Default content to put into an empty cell */
-  emptyCell?: ReactNode;
-  presentationSettings?: Pick<PresentationFieldProps, "borderless" | "typeScale"> &
-    Pick<PresentationContextProps, "wrap">;
-  /** Minimum table width in pixels. Used when calculating columns sizes */
-  minWidthPx?: number;
-  /** Css to apply at each level of a parent/child nested table. */
-  levels?: Record<number, { cellCss?: Properties; firstContentColumn?: Properties }>;
-  /** Allows for customization of the background color used to denote an "active" row */
-  activeBgColor?: Palette;
 }
 
 export interface GridTableDefaults {
@@ -105,12 +51,6 @@ export function setDefaultStyle(style: GridStyle): void {
 export function setGridTableDefaults(opts: Partial<GridTableDefaults>): void {
   defaults = { ...defaults, ...opts };
 }
-
-type RenderAs = "div" | "table" | "virtual";
-
-/** The GridDataRow is optional b/c the nested card chrome rows only have ReactElements. */
-export type RowTuple<R extends Kinded> = [GridDataRow<R> | undefined, ReactElement];
-type ParentChildrenTuple<R extends Kinded> = [GridDataRow<R>, ParentChildrenTuple<R>[]];
 
 /**
  * The sort settings for the current table; whether it's client-side or server-side.
@@ -161,7 +101,7 @@ export interface GridTableProps<R extends Kinded, S, X> {
   /** The rows of data (including any header/footer rows), to be rendered by the column definitions. */
   rows: GridDataRow<R>[];
   /** Optional row-kind-level styling / behavior like onClick/rowLinks. */
-  rowStyles?: GridRowStyles<R>;
+  rowStyles?: RowStyles<R>;
   /** Allow looking up prev/next of a row i.e. for SuperDrawer navigation. */
   rowLookup?: MutableRefObject<GridRowLookup<R> | undefined>;
   /** Whether the header row should be sticky. */
@@ -307,7 +247,7 @@ export function GridTable<R extends Kinded, S = {}, X extends Only<GridTableXss,
       const sortProps = row.kind === "header" ? { sortOn, sortState, setSortKey } : { sortOn };
 
       return (
-        <MemoizedGridRow
+        <Row
           key={`${row.kind}-${row.id}`}
           {...{
             as,
@@ -698,780 +638,3 @@ const VirtualRoot = memoizeOne<(gs: GridStyle, columns: GridColumn<any>[], id: s
     });
   },
 );
-
-/**
- * Calculates column widths using a flexible `calc()` definition that allows for consistent column alignment without the use of `<table />`, CSS Grid, etc layouts.
- * Enforces only fixed-sized units (% and px)
- */
-export function calcColumnSizes(
-  columns: GridColumn<any>[],
-  tableWidth: number | undefined,
-  tableMinWidthPx: number = 0,
-): string[] {
-  // For both default columns (1fr) as well as `w: 4fr` columns, we translate the width into an expression that looks like:
-  // calc((100% - allOtherPercent - allOtherPx) * ((myFr / totalFr))`
-  //
-  // Which looks _a lot_ like how `fr` units just work out-of-the-box.
-  //
-  // Unfortunately, something about having our header & body rows in separate divs (which is controlled
-  // by react-virtuoso), even if they have the same width, for some reason `fr` units between the two
-  // will resolve every slightly differently, where as this approach they will match exactly.
-  const { claimedPercentages, claimedPixels, totalFr } = columns.reduce(
-    (acc, { w }) => {
-      if (typeof w === "undefined") {
-        return { ...acc, totalFr: acc.totalFr + 1 };
-      } else if (typeof w === "number") {
-        return { ...acc, totalFr: acc.totalFr + w };
-      } else if (w.endsWith("fr")) {
-        return { ...acc, totalFr: acc.totalFr + Number(w.replace("fr", "")) };
-      } else if (w.endsWith("px")) {
-        return { ...acc, claimedPixels: acc.claimedPixels + Number(w.replace("px", "")) };
-      } else if (w.endsWith("%")) {
-        return { ...acc, claimedPercentages: acc.claimedPercentages + Number(w.replace("%", "")) };
-      } else {
-        throw new Error("Beam Table column width definition only supports px, percentage, or fr units");
-      }
-    },
-    { claimedPercentages: 0, claimedPixels: 0, totalFr: 0 },
-  );
-
-  // This is our "fake but for some reason it lines up better" fr calc
-  function fr(myFr: number): string {
-    // If the tableWidth, then return a pixel value
-    if (tableWidth) {
-      const widthBasis = Math.max(tableWidth, tableMinWidthPx);
-      return `(${(widthBasis - (claimedPercentages / 100) * widthBasis - claimedPixels) * (myFr / totalFr)}px)`;
-    }
-    // Otherwise return the `calc()` value
-    return `((100% - ${claimedPercentages}% - ${claimedPixels}px) * (${myFr} / ${totalFr}))`;
-  }
-
-  let sizes = columns.map(({ w }) => {
-    if (typeof w === "undefined") {
-      return fr(1);
-    } else if (typeof w === "string") {
-      if (w.endsWith("%") || w.endsWith("px")) {
-        return w;
-      } else if (w.endsWith("fr")) {
-        return fr(Number(w.replace("fr", "")));
-      } else {
-        throw new Error("Beam Table column width definition only supports px, percentage, or fr units");
-      }
-    } else {
-      return fr(w);
-    }
-  });
-
-  return sizes;
-}
-
-/**
- * Given an ADT of type T, performs a look up and returns the type of kind K.
- *
- * See https://stackoverflow.com/a/50125960/355031
- */
-export type DiscriminateUnion<T, K extends keyof T, V extends T[K]> = T extends Record<K, V> ? T : never;
-
-/** A specific kind of row, including the GridDataRow props. */
-type GridRowKind<R extends Kinded, P extends R["kind"]> = DiscriminateUnion<R, "kind", P> & {
-  id: string;
-  children: GridDataRow<R>[];
-  selectable?: false;
-};
-
-/**
- * Defines how a single column will render each given row `kind` in `R`.
- *
- * The `S` generic is either:
- *
- * - For server-side sorting, it's the sortKey to pass back to the server to
- * request "sort by this column".
- *
- * - For client-side sorting, it's type `number`, to represent the current
- * column being sorted, in which case we use the GridCellContent.value.
- */
-export type GridColumn<R extends Kinded, S = {}> = {
-  [K in R["kind"]]:
-    | string
-    | GridCellContent
-    | (DiscriminateUnion<R, "kind", K> extends { data: infer D }
-        ? (
-            data: D,
-            opts: { row: GridRowKind<R, K>; api: GridTableApi<R>; level: number },
-          ) => ReactNode | GridCellContent
-        : (
-            data: undefined,
-            opts: { row: GridRowKind<R, K>; api: GridTableApi<R>; level: number },
-          ) => ReactNode | GridCellContent);
-} & {
-  /**
-   * The column's width.
-   * - Only px, percentage, or fr units are supported, due to each GridRow acting as an independent table. This ensures consistent alignment between rows.
-   * - Numbers are treated as `fr` units
-   * - The default value is `1fr`
-   */
-  w?: number | string;
-  /** The minimum width the column can shrink to */
-  mw?: string;
-  /** The column's default alignment for each cell. */
-  align?: GridCellAlignment;
-  /** Whether the column can be sorted (if client-side sorting). Defaults to true if sorting client-side. */
-  clientSideSort?: boolean;
-  /** This column's sort by value (if server-side sorting). */
-  serverSideSortKey?: S;
-  /** Allows the column to stay in place when the user scrolls horizontally */
-  sticky?: "left" | "right";
-  /** Prevent column from supporting RowStyle.onClick/rowLink in order to avoid nested interactivity. Defaults to true */
-  wrapAction?: false;
-  /** Used as a signal to defer adding the 'indent' styling */
-  isAction?: true;
-  /** Column name that will be used to generate an unique identifier for every row cell */
-  name?: string;
-  /** Flag that will allow to know which columns are hide-able */
-  canHide?: boolean;
-  /** Flag that will allow to know which hide-able columns are visible on initial load */
-  initVisible?: boolean;
-};
-
-export const nonKindGridColumnKeys = [
-  "w",
-  "mw",
-  "align",
-  "serverSideSortKey",
-  "clientSideSort",
-  "sticky",
-  "wrapAction",
-  "isAction",
-];
-
-/** Allows rendering a specific cell. */
-type RenderCellFn<R extends Kinded> = (
-  idx: number,
-  css: Properties,
-  content: ReactNode,
-  row: R,
-  rowStyle: RowStyle<R> | undefined,
-  classNames: string | undefined,
-  onClick: (() => void) | undefined,
-) => ReactNode;
-
-/** Defines row-specific styling for each given row `kind` in `R` */
-export type GridRowStyles<R extends Kinded> = {
-  [P in R["kind"]]?: RowStyle<DiscriminateUnion<R, "kind", P>>;
-};
-
-export interface RowStyle<R extends Kinded> {
-  /** Applies this css to the wrapper row, i.e. for row-level hovers. */
-  rowCss?: Properties | ((row: R) => Properties);
-  /** Applies this css to each cell in the row. */
-  cellCss?: Properties | ((row: R) => Properties);
-  /** Renders the cell element, i.e. a link to get whole-row links. */
-  renderCell?: RenderCellFn<R>;
-  /** Whether the row should be indented (via a style applied to the 1st cell). */
-  indent?: 1 | 2;
-  /** Whether the row should be a link. */
-  rowLink?: (row: R) => string;
-  /** Fired when the row is clicked, similar to rowLink but for actions that aren't 'go to this link'. */
-  onClick?: (row: GridDataRow<R>, api: GridTableApi<R>) => void;
-}
-
-function getIndentationCss<R extends Kinded>(
-  style: GridStyle,
-  rowStyle: RowStyle<R> | undefined,
-  columnIndex: number,
-  maybeContent: ReactNode | GridCellContent,
-): Properties {
-  // Look for cell-specific indent or row-specific indent (row-specific is only one the first column)
-  const indent = (isGridCellContent(maybeContent) && maybeContent.indent) || (columnIndex === 0 && rowStyle?.indent);
-  if (typeof indent === "number" && style.levels !== undefined) {
-    throw new Error(
-      "The indent param is deprecated for new beam fixed & flexible styles, use beamNestedFixedStyle or beamNestedFlexibleStyle",
-    );
-  }
-  return indent === 1 ? style.indentOneCss || {} : indent === 2 ? style.indentTwoCss || {} : {};
-}
-
-function getFirstOrLastCellCss<R extends Kinded>(
-  style: GridStyle,
-  columnIndex: number,
-  columns: GridColumn<R>[],
-): Properties {
-  return {
-    ...(columnIndex === 0 ? style.firstCellCss : {}),
-    ...(columnIndex === columns.length - 1 ? style.lastCellCss : {}),
-  };
-}
-
-export type GridCellAlignment = "left" | "right" | "center";
-
-/**
- * Allows a cell to be more than just a RectNode, i.e. declare its alignment or
- * primitive value for filtering and sorting.
- */
-export type GridCellContent = {
-  /** The JSX content of the cell. Virtual tables that client-side sort should use a function to avoid perf overhead. */
-  content: ReactNode | (() => ReactNode);
-  alignment?: GridCellAlignment;
-  /** Allow value to be a function in case it's a dynamic value i.e. reading from an inline-edited proxy. */
-  value?: MaybeFn<number | string | Date | boolean | null | undefined>;
-  /** The value to use specifically for sorting (i.e. if `value` is used for filtering); defaults to `value`. */
-  sortValue?: MaybeFn<number | string | Date | boolean | null | undefined>;
-  /** Whether to indent the cell. */
-  indent?: 1 | 2;
-  colspan?: number;
-  typeScale?: Typography;
-  /** Allows the cell to stay in place when the user scrolls horizontally, i.e. frozen columns. */
-  sticky?: "left" | "right";
-  /** If provided, content of the cell will be wrapped within a <button /> or <a /> tag depending on if the value is a function or a string. */
-  onClick?: () => {} | string;
-  /** Custom css to apply directly to this cell, i.e. cell-specific borders. */
-  css?: Properties;
-  /** Allows cell to reveal content when the user hovers over a row. Content must be wrapped in an element in order to be hidden. IE <div>{value}</div>*/
-  revealOnRowHover?: true;
-};
-
-type MaybeFn<T> = T | (() => T);
-
-/**
- * The data for any row in the table, marked by `kind` so that each column knows how to render it.
- *
- * Each `kind` should contain very little presentation logic, i.e. mostly just off-the-wire data from
- * a GraphQL query.
- *
- * The presentation concerns instead mainly live in each GridColumn definition, which will format/render
- * each kind's data for that specific row+column (i.e. cell) combination.
- */
-export type GridDataRow<R extends Kinded> = {
-  kind: R["kind"];
-  /** Combined with the `kind` to determine a table wide React key. */
-  id: string;
-  /** A list of parent/grand-parent ids for collapsing parent/child rows. */
-  children?: GridDataRow<R>[];
-  /** * Whether to pin this sort to the first/last of its parent's children.
-   *
-   * By default, pinned rows are always shown/not filtered out, however providing
-   * the pin `filter: true` property will allow pinned rows to be hidden
-   * while filtering.*/
-  pin?: "first" | "last" | Pin;
-  data: unknown;
-  /** Whether to have the row collapsed (children not visible) on initial load. This will be ignore in subsequent re-renders of the table */
-  initCollapsed?: boolean;
-  /** Whether to have the row selected on initial load. This will be ignore in subsequent re-renders of the table */
-  initSelected?: boolean;
-  /** Whether row can be selected */
-  selectable?: false;
-} & IfAny<R, {}, DiscriminateUnion<R, "kind", R["kind"]>>;
-
-/**
- * Used to indicate where to pin the DataRow and if whether it should be filtered or always visible, setting `filter` to `true` will hide this row
- * if it doesn't match the provided filtering search term
- */
-export type Pin = { at: "first" | "last"; filter?: boolean };
-
-// Use IfAny so that GridDataRow<any> doesn't devolve into any
-type IfAny<T, Y, N> = 0 extends 1 & T ? Y : N;
-
-interface GridRowProps<R extends Kinded, S> {
-  as: RenderAs;
-  columns: GridColumn<R>[];
-  row: GridDataRow<R>;
-  style: GridStyle;
-  rowStyles: GridRowStyles<R> | undefined;
-  stickyHeader: boolean;
-  stickyOffset: number;
-  sortOn: SortOn;
-  sortState?: SortState<S>;
-  setSortKey?: (value: S) => void;
-  columnSizes: string[];
-  level: number;
-  getCount: (id: string) => object;
-  api: GridTableApi<R>;
-  cellHighlight: boolean;
-}
-
-// We extract GridRow to its own mini-component primarily so we can React.memo'ize it.
-function GridRow<R extends Kinded, S>(props: GridRowProps<R, S>): ReactElement {
-  const {
-    as,
-    columns,
-    row,
-    style,
-    rowStyles,
-    stickyHeader,
-    stickyOffset,
-    sortOn,
-    sortState,
-    setSortKey,
-    columnSizes,
-    level,
-    getCount,
-    api,
-    cellHighlight,
-    ...others
-  } = props;
-
-  const { rowState } = useContext(RowStateContext);
-  const rowId = `${row.kind}_${row.id}`;
-  const isActive = useComputed(() => rowState.activeRowId === rowId, [rowId, rowState]);
-
-  // We treat the "header" and "totals" kind as special for "good defaults" styling
-  const isHeader = row.kind === "header";
-  const isTotals = row.kind === "totals";
-  const rowStyle = rowStyles?.[row.kind];
-  const Row = as === "table" ? "tr" : "div";
-
-  const revealOnRowHoverClass = "revealOnRowHover";
-
-  const rowStyleCellCss = maybeApplyFunction(row as any, rowStyle?.cellCss);
-  const rowCss = {
-    // For virtual tables use `display: flex` to keep all cells on the same row. For each cell in the row use `flexNone` to ensure they stay their defined widths
-    ...(as === "table" ? {} : Css.relative.df.fg1.fs1.addIn("&>*", Css.flexNone.$).$),
-    ...((rowStyle?.rowLink || rowStyle?.onClick) && {
-      // Even though backgroundColor is set on the cellCss, the hover target is the row.
-      "&:hover > *": Css.cursorPointer.bgColor(style.rowHoverColor ?? Palette.LightBlue100).$,
-    }),
-    ...maybeApplyFunction(row as any, rowStyle?.rowCss),
-    // Maybe add the sticky header styles
-    ...((isHeader || isTotals) && stickyHeader ? Css.sticky.topPx(stickyOffset).z2.$ : undefined),
-    ...{
-      [` > .${revealOnRowHoverClass} > *`]: Css.invisible.$,
-      [`:hover > .${revealOnRowHoverClass} > *`]: Css.visible.$,
-    },
-  };
-
-  let currentColspan = 1;
-
-  let firstContentColumnStylesApplied = false;
-
-  return (
-    <Row css={rowCss} {...others} data-gridrow {...getCount(row.id)}>
-      {columns.map((column, columnIndex) => {
-        const { wrapAction = true, isAction = false } = column;
-
-        const applyFirstContentColumnStyles = !isHeader && !isAction && !firstContentColumnStylesApplied;
-        firstContentColumnStylesApplied ||= applyFirstContentColumnStyles;
-
-        if (column.mw) {
-          // Validate the column's minWidth definition if set.
-          if (!column.mw.endsWith("px") && !column.mw.endsWith("%")) {
-            throw new Error("Beam Table column min-width definition only supports px or percentage values");
-          }
-        }
-
-        // Decrement colspan count and skip if greater than 1.
-        if (currentColspan > 1) {
-          currentColspan -= 1;
-          return null;
-        }
-        const maybeContent = applyRowFn(column, row, api, level);
-        currentColspan = isGridCellContent(maybeContent) ? maybeContent.colspan ?? 1 : 1;
-        const revealOnRowHover = isGridCellContent(maybeContent) ? maybeContent.revealOnRowHover : false;
-
-        const canSortColumn =
-          (sortOn === "client" && column.clientSideSort !== false) ||
-          (sortOn === "server" && !!column.serverSideSortKey);
-        const alignment = getAlignment(column, maybeContent);
-        const justificationCss = getJustification(column, maybeContent, as, alignment);
-        const content = toContent(maybeContent, isHeader, canSortColumn, sortOn === "client", style, as, alignment);
-
-        ensureClientSideSortValueIsSortable(sortOn, isHeader, column, columnIndex, maybeContent);
-
-        const maybeSticky = ((isGridCellContent(maybeContent) && maybeContent.sticky) || column.sticky) ?? undefined;
-        const maybeStickyColumnStyles =
-          maybeSticky && columnSizes
-            ? {
-                ...Css.sticky.z1.bgWhite.$,
-                ...(maybeSticky === "left"
-                  ? Css.left(columnIndex === 0 ? 0 : `calc(${columnSizes.slice(0, columnIndex).join(" + ")})`).$
-                  : {}),
-                ...(maybeSticky === "right"
-                  ? Css.right(
-                      columnIndex + 1 === columnSizes.length
-                        ? 0
-                        : `calc(${columnSizes.slice(columnIndex + 1 - columnSizes.length).join(" + ")})`,
-                    ).$
-                  : {}),
-              }
-            : {};
-
-        const cellId = `${row.kind}_${row.id}_${column.name}`;
-        const applyCellHighlight = cellHighlight && !!column.name && !isHeader && !isTotals;
-        const isCellActive = rowState.activeCellId === cellId;
-
-        // Note that it seems expensive to calc a per-cell class name/CSS-in-JS output,
-        // vs. setting global/table-wide CSS like `style.cellCss` on the root grid div with
-        // a few descendent selectors. However, that approach means the root grid-applied
-        // CSS has a high-specificity and so its harder for per-page/per-cell business logic
-        // to override it. So, we just calc the combined table-wide+per-cell-overridden CSS here,
-        // in a very CSS-in-JS idiomatic manner.
-        //
-        // In practice we've not seen any performance issues with this from our "large but
-        // not Google spreadsheets" tables.
-        const cellCss = {
-          // Adding display flex so we can align content within the cells
-          ...Css.df.$,
-          // Apply sticky column/cell styles
-          ...maybeStickyColumnStyles,
-          // Apply any static/all-cell styling
-          ...style.cellCss,
-          // Then override with first/last cell styling
-          ...getFirstOrLastCellCss(style, columnIndex, columns),
-          // Then override with per-cell/per-row justification/indentation
-          ...justificationCss,
-          ...getIndentationCss(style, rowStyle, columnIndex, maybeContent),
-          // Then apply any header-specific override
-          ...(isHeader && style.headerCellCss),
-          // Then apply any totals-specific override
-          ...(isTotals && style.totalsCellCss),
-          // Or level-specific styling
-          ...(!isHeader && !isTotals && !!style.levels && style.levels[level]?.cellCss),
-          // Level specific styling for the first content column
-          ...(applyFirstContentColumnStyles && !!style.levels && style.levels[level]?.firstContentColumn),
-          // The specific cell's css (if any from GridCellContent)
-          ...rowStyleCellCss,
-          // Apply active row styling for non-nested card styles.
-          ...(isActive ? Css.bgColor(style.activeBgColor ?? Palette.LightBlue50).$ : {}),
-          // Add any cell specific style overrides
-          ...(isGridCellContent(maybeContent) && maybeContent.typeScale ? Css[maybeContent.typeScale].$ : {}),
-          // And any cell specific css
-          ...(isGridCellContent(maybeContent) && maybeContent.css ? maybeContent.css : {}),
-          // Apply cell highlight styles to active cell and hover
-          ...Css.if(applyCellHighlight && isCellActive).br4.boxShadow(`inset 0 0 0 1px ${Palette.LightBlue700}`).$,
-          // Define the width of the column on each cell. Supports col spans.
-          width: `calc(${columnSizes.slice(columnIndex, columnIndex + currentColspan).join(" + ")})`,
-          ...(column.mw ? Css.mw(column.mw).$ : {}),
-        };
-
-        const cellClassNames = revealOnRowHover ? revealOnRowHoverClass : undefined;
-
-        const cellOnClick = applyCellHighlight ? () => api.setActiveCellId(cellId) : undefined;
-
-        const renderFn: RenderCellFn<any> =
-          (rowStyle?.renderCell || rowStyle?.rowLink) && wrapAction
-            ? rowLinkRenderFn(as)
-            : isHeader
-            ? headerRenderFn(columns, column, sortState, setSortKey, as)
-            : rowStyle?.onClick && wrapAction
-            ? rowClickRenderFn(as, api)
-            : defaultRenderFn(as);
-
-        return renderFn(columnIndex, cellCss, content, row, rowStyle, cellClassNames, cellOnClick);
-      })}
-    </Row>
-  );
-}
-
-/**
- * Memoizes GridRows so that re-rendering the table doesn't re-render every single row.
- *
- * We use a custom `propsAreEqual` for the `GridRowProps.row` property, which we memoize
- * based on the `GridDataRow.data` prop, such that if a table re-renders (i.e. for a cache
- * updated) and technically creates new row instances, but a row's `row.data` points to the
- * same/unchanged Apollo cache fragment, then we won't re-render that row.
- *
- * Note that if you're using virtualization, it can be surprising how unnoticeable broken row
- * memoization is.
- */
-// Declared as a const + `as typeof GridRow` to work with generics, see https://github.com/DefinitelyTyped/DefinitelyTyped/issues/37087#issuecomment-656596623
-const MemoizedGridRow = React.memo(observer(GridRow), (one, two) => {
-  const { row: row1, ...others1 } = one;
-  const { row: row2, ...others2 } = two;
-  return shallowEqual(row1, row2) && shallowEqual(others1, others2);
-}) as typeof GridRow;
-
-/** A heuristic to detect the result of `React.createElement` / i.e. JSX. */
-function isJSX(content: any): boolean {
-  return typeof content === "object" && content && "type" in content && "props" in content;
-}
-
-/** If a column def return just string text for a given row, apply some default styling. */
-function toContent(
-  maybeContent: ReactNode | GridCellContent,
-  isHeader: boolean,
-  canSortColumn: boolean,
-  isClientSideSorting: boolean,
-  style: GridStyle,
-  as: RenderAs,
-  alignment: GridCellAlignment,
-): ReactNode {
-  let content = isGridCellContent(maybeContent) ? maybeContent.content : maybeContent;
-  if (typeof content === "function") {
-    // Actually create the JSX by calling `content()` here (which should be as late as
-    // possible, i.e. only for visible rows if we're in a virtual table).
-    content = content();
-  } else if (as === "virtual" && canSortColumn && isClientSideSorting && isJSX(content)) {
-    // When using client-side sorting, we call `applyRowFn` not only during rendering, but
-    // up-front against all rows (for the currently sorted column) to determine their
-    // sort values.
-    //
-    // Pedantically this means that any table using client-side sorting should not
-    // build JSX directly in its GridColumn functions, but this overhead is especially
-    // noticeable for large/virtualized tables, so we only enforce using functions
-    // for those tables.
-    throw new Error(
-      "GridTables with as=virtual & sortable columns should use functions that return JSX, instead of JSX",
-    );
-  }
-
-  content =
-    isGridCellContent(maybeContent) && !!maybeContent.onClick
-      ? getButtonOrLink(content, maybeContent.onClick, {
-          css: Css.maxw100.lightBlue700.ta("inherit").if(style?.presentationSettings?.wrap === false).truncate.$,
-        })
-      : content;
-
-  if (content && typeof content === "string" && isHeader && canSortColumn) {
-    return <SortHeader content={content} iconOnLeft={alignment === "right"} />;
-  } else if (content && style?.presentationSettings?.wrap === false && typeof content === "string") {
-    // In order to truncate the text properly, then we need to wrap it in another element
-    // as our cell element is a flex container, which don't allow for applying truncation styles directly on it.
-    return (
-      <span css={Css.truncate.mw0.$} title={content}>
-        {content}
-      </span>
-    );
-  } else if (style.emptyCell && isContentEmpty(content)) {
-    // If the content is empty and the user specified an `emptyCell` node, return that.
-    return style.emptyCell;
-  }
-  return content;
-}
-
-function isGridCellContent(content: ReactNode | GridCellContent): content is GridCellContent {
-  return typeof content === "object" && !!content && "content" in content;
-}
-
-const emptyValues = ["", null, undefined] as any[];
-function isContentEmpty(content: ReactNode): boolean {
-  return emptyValues.includes(content);
-}
-
-/** Return the content for a given column def applied to a given row. */
-export function applyRowFn<R extends Kinded>(
-  column: GridColumn<R>,
-  row: GridDataRow<R>,
-  api: GridTableApi<R>,
-  level: number,
-): ReactNode | GridCellContent {
-  // Usually this is a function to apply against the row, but sometimes it's a hard-coded value, i.e. for headers
-  const maybeContent = column[row.kind];
-  if (typeof maybeContent === "function") {
-    // Auto-destructure data
-    return (maybeContent as Function)((row as any)["data"], { row: row as any, api, level });
-  } else {
-    return maybeContent;
-  }
-}
-
-/** Renders our default cell element, i.e. if no row links and no custom renderCell are used. */
-const defaultRenderFn: (as: RenderAs) => RenderCellFn<any> =
-  (as: RenderAs) => (key, css, content, row, rowStyle, classNames: string | undefined, onClick) => {
-    const Cell = as === "table" ? "td" : "div";
-    return (
-      <Cell key={key} css={{ ...css, ...tableRowStyles(as) }} className={classNames} onClick={onClick}>
-        {content}
-      </Cell>
-    );
-  };
-
-/** Sets up the `GridContext` so that header cells can access the current sort settings. */
-const headerRenderFn: (
-  columns: GridColumn<any>[],
-  column: GridColumn<any>,
-  sortState: SortState<any> | undefined,
-  setSortKey: Function | undefined,
-  as: RenderAs,
-) => RenderCellFn<any> =
-  (columns, column, sortState, setSortKey, as) =>
-  (key, css, content, row, rowStyle, classNames: string | undefined) => {
-    const [currentKey, direction] = sortState || [];
-    // If server-side sorting, use the user's key for this column; client-side sorting, use the index.
-    const ourSortKey = column.serverSideSortKey || columns.indexOf(column);
-    const context: GridSortContextProps = {
-      sorted: ourSortKey === currentKey ? direction : undefined,
-      toggleSort: () => setSortKey!(ourSortKey),
-    };
-    const Cell = as === "table" ? "th" : "div";
-    return (
-      <GridSortContext.Provider key={key} value={context}>
-        <Cell css={{ ...css, ...tableRowStyles(as, column) }} className={classNames}>
-          {content}
-        </Cell>
-      </GridSortContext.Provider>
-    );
-  };
-
-/** Renders a cell element when a row link is in play. */
-const rowLinkRenderFn: (as: RenderAs) => RenderCellFn<any> =
-  (as: RenderAs) => (key, css, content, row, rowStyle, classNames: string | undefined) => {
-    const to = rowStyle!.rowLink!(row);
-    if (as === "table") {
-      return (
-        <td key={key} css={{ ...css, ...tableRowStyles(as) }} className={classNames}>
-          <Link to={to} css={Css.noUnderline.color("unset").db.$} className={navLink}>
-            {content}
-          </Link>
-        </td>
-      );
-    }
-    return (
-      <Link
-        key={key}
-        to={to}
-        css={{ ...Css.noUnderline.color("unset").$, ...css }}
-        className={`${navLink} ${classNames}`}
-      >
-        {content}
-      </Link>
-    );
-  };
-
-/** Renders a cell that will fire the RowStyle.onClick. */
-const rowClickRenderFn: (as: RenderAs, api: GridTableApi<any>) => RenderCellFn<any> =
-  (as: RenderAs, api: GridTableApi<any>) =>
-  (key, css, content, row, rowStyle, classNames: string | undefined, onClick) => {
-    const Row = as === "table" ? "tr" : "div";
-    return (
-      <Row
-        {...{ key }}
-        css={{ ...css, ...tableRowStyles(as) }}
-        className={classNames}
-        onClick={(e) => {
-          rowStyle!.onClick!(row, api);
-          onClick && onClick();
-        }}
-      >
-        {content}
-      </Row>
-    );
-  };
-
-const alignmentToJustify: Record<GridCellAlignment, Properties["justifyContent"]> = {
-  left: "flex-start",
-  center: "center",
-  right: "flex-end",
-};
-const alignmentToTextAlign: Record<GridCellAlignment, Properties["textAlign"]> = {
-  left: "left",
-  center: "center",
-  right: "right",
-};
-
-function getAlignment(column: GridColumn<any>, maybeContent: ReactNode | GridCellContent): GridCellAlignment {
-  return (isGridCellContent(maybeContent) && maybeContent.alignment) || column.align || "left";
-}
-
-// For alignment, use: 1) cell def, else 2) column def, else 3) left.
-function getJustification(
-  column: GridColumn<any>,
-  maybeContent: ReactNode | GridCellContent,
-  as: RenderAs,
-  alignment: GridCellAlignment,
-) {
-  // Always apply text alignment.
-  const textAlign = Css.add("textAlign", alignmentToTextAlign[alignment]).$;
-  if (as === "table") {
-    return textAlign;
-  }
-  return { ...Css.jc(alignmentToJustify[alignment]).$, ...textAlign };
-}
-
-/** Look at a row and get its filter value. */
-function filterValue(value: ReactNode | GridCellContent): any {
-  let maybeFn = value;
-  if (value && typeof value === "object") {
-    if ("value" in value) {
-      maybeFn = value.value;
-    } else if ("content" in value) {
-      maybeFn = value.content;
-    }
-  }
-  // Watch for functions that need to read from a potentially-changing proxy
-  if (maybeFn instanceof Function) {
-    return maybeFn();
-  }
-  return maybeFn;
-}
-
-function maybeApplyFunction<T>(
-  row: T,
-  maybeFn: Properties | ((row: T) => Properties) | undefined,
-): Properties | undefined {
-  return typeof maybeFn === "function" ? maybeFn(row) : maybeFn;
-}
-
-export function matchesFilter(maybeContent: ReactNode | GridCellContent, filter: string): boolean {
-  let value = filterValue(maybeContent);
-  if (typeof value === "string") {
-    return value.toLowerCase().includes(filter.toLowerCase());
-  } else if (typeof value === "number") {
-    return Number(filter) === value;
-  }
-  return false;
-}
-
-/** GridTable as Table utility to apply <tr> element override styles. */
-function tableRowStyles(as: RenderAs, column?: GridColumn<any>) {
-  const thWidth = column?.w;
-  return as === "table"
-    ? {
-        ...Css.dtc.$,
-        ...(thWidth ? Css.w(thWidth).$ : {}),
-      }
-    : {};
-}
-
-function resolveStyles(style: GridStyle | GridStyleDef): GridStyle {
-  const defKeys: (keyof GridStyleDef)[] = ["inlineEditing", "grouped", "rowHeight", "cellHighlight"];
-  const keys = safeKeys(style);
-  if (keys.length === 0 || keys.some((k) => defKeys.includes(k))) {
-    return getTableStyles(style as GridStyleDef);
-  }
-  return style as GridStyle;
-}
-
-/**
- * Filters rows given a client-side text `filter.
- *
- * Ensures parent rows remain in the list if any children match the filter.
- *
- * We return a copy of `[Parent, [Child]]` tuples so that we don't modify the `GridDataRow.children`.
- */
-function filterRows<R extends Kinded>(
-  api: GridTableApi<R>,
-  columns: GridColumn<R>[],
-  rows: GridDataRow<R>[],
-  filter: string | undefined,
-): ParentChildrenTuple<R>[] {
-  // Make a functions to do recursion
-  function acceptAll(acc: ParentChildrenTuple<R>[], row: GridDataRow<R>): ParentChildrenTuple<R>[] {
-    return acc.concat([[row, row.children?.reduce(acceptAll, []) ?? []]]);
-  }
-
-  function filterFn(acc: ParentChildrenTuple<R>[], row: GridDataRow<R>): ParentChildrenTuple<R>[] {
-    // Break up "foo bar" into `[foo, bar]` and a row must match both `foo` and `bar`
-    const filters = (filter && filter.split(/ +/)) || [];
-    const matches =
-      row.kind === "header" ||
-      row.kind === "totals" ||
-      filters.length === 0 ||
-      filters.every((f) =>
-        columns.map((c) => applyRowFn(c, row, api, 0)).some((maybeContent) => matchesFilter(maybeContent, f)),
-      );
-    if (matches) {
-      return acc.concat([[row, row.children?.reduce(acceptAll, []) ?? []]]);
-    } else {
-      const matchedChildren = row.children?.reduce(filterFn, []) ?? [];
-      if (
-        matchedChildren.length > 0 ||
-        typeof row.pin === "string" ||
-        (row.pin !== undefined && row.pin.filter !== true)
-      ) {
-        return acc.concat([[row, matchedChildren]]);
-      } else {
-        return acc;
-      }
-    }
-  }
-  return rows.reduce(filterFn, []);
-}
