+++ src/components/Table/Table.qa.stories.tsx
@@ -3,11 +3,13 @@ import { Meta } from "@storybook/react";
 import { default as React, ReactNode, useMemo, useState } from "react";
 import { Chips } from "src/components/Chips";
 import { Icon } from "src/components/Icon";
-import { collapseColumn, column, dateColumn, numericColumn, selectColumn } from "src/components/Table/columns";
-import { emptyCell, GridColumn, GridDataRow, GridSortConfig, GridTable } from "src/components/Table/GridTable";
+import { GridSortConfig, GridTable } from "src/components/Table/GridTable";
 import { useGridTableApi } from "src/components/Table/GridTableApi";
-import { simpleHeader, SimpleHeaderAndData } from "src/components/Table/simpleHelpers";
-import { getTableStyles } from "src/components/Table/styles";
+import { GridColumn, GridDataRow } from "src/components/Table/types";
+import { emptyCell } from "src/components/Table/utils";
+import { collapseColumn, column, dateColumn, numericColumn, selectColumn } from "src/components/Table/utils/columns";
+import { simpleHeader, SimpleHeaderAndData } from "src/components/Table/utils/simpleHelpers";
+import { getTableStyles } from "src/components/Table/utils/styles";
 import { Tag } from "src/components/Tag";
 import { Css, Palette } from "src/Css";
 import {
