+++ src/components/Table/hooks/index.ts
@@ -1,0 +1,3 @@
+export * from "./useColumns";
+export * from "./useSetupColumnSizes";
+export * from "./useSortState";
