+++ src/components/Table/hooks/useSortState.test.ts
@@ -1,5 +1,5 @@
-import { ASC, DESC } from "src/components/Table/GridTable";
-import { deriveSortState } from "src/components/Table/useSortState";
+import { deriveSortState } from "src/components/Table/hooks/useSortState";
+import { ASC, DESC } from "src/components/Table/utils";
 
 describe("useSortState", () => {
   it("can derive the next sort state when current state is undefined", () => {
@@ -13,15 +13,30 @@ describe("useSortState", () => {
   it("can derive the next sort state when clicking on a currently ascending column", () => {
     expect(deriveSortState([1, ASC, undefined, undefined], 1, undefined)).toEqual([1, DESC, undefined, undefined]);
     // With `initialSortState` the same as the currentSortState Ascending
-    expect(deriveSortState([1, ASC, undefined, undefined], 1, [1, ASC, undefined, undefined])).toEqual([1, DESC, undefined, undefined]);
+    expect(deriveSortState([1, ASC, undefined, undefined], 1, [1, ASC, undefined, undefined])).toEqual([
+      1,
+      DESC,
+      undefined,
+      undefined,
+    ]);
   });
 
   it("can derive the next sort state when clicking on a currently descending column", () => {
     // With `initialSortState` defined
-    expect(deriveSortState([1, DESC, undefined, undefined], 1, [2, ASC, undefined, undefined])).toEqual([2, ASC, undefined, undefined]);
+    expect(deriveSortState([1, DESC, undefined, undefined], 1, [2, ASC, undefined, undefined])).toEqual([
+      2,
+      ASC,
+      undefined,
+      undefined,
+    ]);
     // Without `initialSortState` defined
     expect(deriveSortState([1, DESC, undefined, undefined], 1, undefined)).toEqual(undefined);
     // With `initialSortState` the same as the currentSortState Descending
-    expect(deriveSortState([1, DESC, undefined, undefined], 1, [1, DESC, undefined, undefined])).toEqual([1, ASC, undefined, undefined]);
+    expect(deriveSortState([1, DESC, undefined, undefined], 1, [1, DESC, undefined, undefined])).toEqual([
+      1,
+      ASC,
+      undefined,
+      undefined,
+    ]);
   });
 });
