+++ src/components/Table/index.ts
@@ -1,31 +1,17 @@
-export * from "./CollapseToggle";
-export * from "./columns";
-export { EditColumnsButton } from "./EditColumnsButton";
-export type { GridRowLookup } from "./GridRowLookup";
-export { GridSortContext } from "./GridSortContext";
-export { ASC, DESC, GridTable, setDefaultStyle, setGridTableDefaults } from "./GridTable";
-export type {
-  Direction,
-  GridCellAlignment,
-  GridCellContent,
-  GridColumn,
-  GridDataRow,
-  GridRowStyles,
-  GridSortConfig,
-  GridStyle,
-  GridTableDefaults,
-  GridTableProps,
-  GridTableXss,
-  Kinded,
-  RowStyle,
-  setRunningInJest,
-} from "./GridTable";
+export * from "src/components/Table/components/SelectToggle";
+export { SortHeader } from "src/components/Table/components/SortHeader";
+export { useColumns } from "src/components/Table/hooks/useColumns";
+export * from "src/components/Table/utils/columns";
+export type { GridRowLookup } from "src/components/Table/utils/GridRowLookup";
+export { GridSortContext } from "src/components/Table/utils/GridSortContext";
+export { RowState, RowStateContext } from "src/components/Table/utils/RowState";
+export { simpleDataRows, simpleHeader } from "src/components/Table/utils/simpleHelpers";
+export type { SimpleHeaderAndData } from "src/components/Table/utils/simpleHelpers";
+export * from "./components";
+export { GridTable, setDefaultStyle, setGridTableDefaults } from "./GridTable";
+export type { GridSortConfig, GridTableDefaults, GridTableProps, setRunningInJest } from "./GridTable";
 export { useGridTableApi } from "./GridTableApi";
 export type { GridTableApi } from "./GridTableApi";
-export { RowState, RowStateContext } from "./RowState";
-export * from "./SelectToggle";
-export { simpleDataRows, simpleHeader } from "./simpleHelpers";
-export type { SimpleHeaderAndData } from "./simpleHelpers";
-export { SortHeader } from "./SortHeader";
-export { cardStyle, condensedStyle, defaultStyle, getTableStyles } from "./styles";
-export { useColumns } from "./useColumns";
+export * from "./hooks";
+export * from "./types";
+export * from "./utils";
