+++ src/components/Table/utils/sortRows.test.ts
@@ -1,6 +1,6 @@
-import { GridColumn, GridDataRow } from "src/components/Table/GridTable";
-import { simpleHeader } from "src/components/Table/simpleHelpers";
-import { sortRows } from "src/components/Table/sortRows";
+import { GridColumn, GridDataRow } from "src/components/Table/types";
+import { simpleHeader } from "src/components/Table/utils/simpleHelpers";
+import { sortRows } from "src/components/Table/utils/sortRows";
 
 describe("sortRows", () => {
   it("can return a sorted copy of nested rows in ascending order", () => {
