+++ src/components/Table/utils/utils.tsx
@@ -1,0 +1,240 @@
+import React, { ReactNode } from "react";
+import { GridCellContent } from "src/components/Table/components";
+import { SortHeader } from "src/components/Table/components/SortHeader";
+import { GridTableApi } from "src/components/Table/GridTableApi";
+import {
+  GridCellAlignment,
+  GridColumn,
+  GridDataRow,
+  Kinded,
+  ParentChildrenTuple,
+  RenderAs,
+} from "src/components/Table/types";
+import { GridStyle, RowStyle } from "src/components/Table/utils/styles";
+import { Css, Properties } from "src/Css";
+import { getButtonOrLink } from "src/utils/getInteractiveElement";
+
+/** If a column def return just string text for a given row, apply some default styling. */
+export function toContent(
+  maybeContent: ReactNode | GridCellContent,
+  isHeader: boolean,
+  canSortColumn: boolean,
+  isClientSideSorting: boolean,
+  style: GridStyle,
+  as: RenderAs,
+  alignment: GridCellAlignment,
+): ReactNode {
+  let content = isGridCellContent(maybeContent) ? maybeContent.content : maybeContent;
+  if (typeof content === "function") {
+    // Actually create the JSX by calling `content()` here (which should be as late as
+    // possible, i.e. only for visible rows if we're in a virtual table).
+    content = content();
+  } else if (as === "virtual" && canSortColumn && isClientSideSorting && isJSX(content)) {
+    // When using client-side sorting, we call `applyRowFn` not only during rendering, but
+    // up-front against all rows (for the currently sorted column) to determine their
+    // sort values.
+    //
+    // Pedantically this means that any table using client-side sorting should not
+    // build JSX directly in its GridColumn functions, but this overhead is especially
+    // noticeable for large/virtualized tables, so we only enforce using functions
+    // for those tables.
+    throw new Error(
+      "GridTables with as=virtual & sortable columns should use functions that return JSX, instead of JSX",
+    );
+  }
+
+  content =
+    isGridCellContent(maybeContent) && !!maybeContent.onClick
+      ? getButtonOrLink(content, maybeContent.onClick, {
+          css: Css.maxw100.lightBlue700.ta("inherit").if(style?.presentationSettings?.wrap === false).truncate.$,
+        })
+      : content;
+
+  if (content && typeof content === "string" && isHeader && canSortColumn) {
+    return <SortHeader content={content} iconOnLeft={alignment === "right"} />;
+  } else if (content && style?.presentationSettings?.wrap === false && typeof content === "string") {
+    // In order to truncate the text properly, then we need to wrap it in another element
+    // as our cell element is a flex container, which don't allow for applying truncation styles directly on it.
+    return (
+      <span css={Css.truncate.mw0.$} title={content}>
+        {content}
+      </span>
+    );
+  } else if (style.emptyCell && isContentEmpty(content)) {
+    // If the content is empty and the user specified an `emptyCell` node, return that.
+    return style.emptyCell;
+  }
+  return content;
+}
+
+export function isGridCellContent(content: ReactNode | GridCellContent): content is GridCellContent {
+  return typeof content === "object" && !!content && "content" in content;
+}
+
+const emptyValues = ["", null, undefined] as any[];
+function isContentEmpty(content: ReactNode): boolean {
+  return emptyValues.includes(content);
+}
+
+/** Return the content for a given column def applied to a given row. */
+export function applyRowFn<R extends Kinded>(
+  column: GridColumn<R>,
+  row: GridDataRow<R>,
+  api: GridTableApi<R>,
+  level: number,
+): ReactNode | GridCellContent {
+  // Usually this is a function to apply against the row, but sometimes it's a hard-coded value, i.e. for headers
+  const maybeContent = column[row.kind];
+  if (typeof maybeContent === "function") {
+    // Auto-destructure data
+    return (maybeContent as Function)((row as any)["data"], { row: row as any, api, level });
+  } else {
+    return maybeContent;
+  }
+}
+
+/**
+ * Filters rows given a client-side text `filter.
+ *
+ * Ensures parent rows remain in the list if any children match the filter.
+ *
+ * We return a copy of `[Parent, [Child]]` tuples so that we don't modify the `GridDataRow.children`.
+ */
+export function filterRows<R extends Kinded>(
+  api: GridTableApi<R>,
+  columns: GridColumn<R>[],
+  rows: GridDataRow<R>[],
+  filter: string | undefined,
+): ParentChildrenTuple<R>[] {
+  // Make a functions to do recursion
+  function acceptAll(acc: ParentChildrenTuple<R>[], row: GridDataRow<R>): ParentChildrenTuple<R>[] {
+    return acc.concat([[row, row.children?.reduce(acceptAll, []) ?? []]]);
+  }
+
+  function filterFn(acc: ParentChildrenTuple<R>[], row: GridDataRow<R>): ParentChildrenTuple<R>[] {
+    // Break up "foo bar" into `[foo, bar]` and a row must match both `foo` and `bar`
+    const filters = (filter && filter.split(/ +/)) || [];
+    const matches =
+      row.kind === "header" ||
+      row.kind === "totals" ||
+      filters.length === 0 ||
+      filters.every((f) =>
+        columns.map((c) => applyRowFn(c, row, api, 0)).some((maybeContent) => matchesFilter(maybeContent, f)),
+      );
+    if (matches) {
+      return acc.concat([[row, row.children?.reduce(acceptAll, []) ?? []]]);
+    } else {
+      const matchedChildren = row.children?.reduce(filterFn, []) ?? [];
+      if (
+        matchedChildren.length > 0 ||
+        typeof row.pin === "string" ||
+        (row.pin !== undefined && row.pin.filter !== true)
+      ) {
+        return acc.concat([[row, matchedChildren]]);
+      } else {
+        return acc;
+      }
+    }
+  }
+  return rows.reduce(filterFn, []);
+}
+
+export const ASC = "ASC" as const;
+export const DESC = "DESC" as const;
+export const emptyCell: GridCellContent = { content: () => <></>, value: "" };
+
+export function getIndentationCss<R extends Kinded>(
+  style: GridStyle,
+  rowStyle: RowStyle<R> | undefined,
+  columnIndex: number,
+  maybeContent: ReactNode | GridCellContent,
+): Properties {
+  // Look for cell-specific indent or row-specific indent (row-specific is only one the first column)
+  const indent = (isGridCellContent(maybeContent) && maybeContent.indent) || (columnIndex === 0 && rowStyle?.indent);
+  if (typeof indent === "number" && style.levels !== undefined) {
+    throw new Error(
+      "The indent param is deprecated for new beam fixed & flexible styles, use beamNestedFixedStyle or beamNestedFlexibleStyle",
+    );
+  }
+  return indent === 1 ? style.indentOneCss || {} : indent === 2 ? style.indentTwoCss || {} : {};
+}
+
+export function getFirstOrLastCellCss<R extends Kinded>(
+  style: GridStyle,
+  columnIndex: number,
+  columns: GridColumn<R>[],
+): Properties {
+  return {
+    ...(columnIndex === 0 ? style.firstCellCss : {}),
+    ...(columnIndex === columns.length - 1 ? style.lastCellCss : {}),
+  };
+}
+
+/** A heuristic to detect the result of `React.createElement` / i.e. JSX. */
+function isJSX(content: any): boolean {
+  return typeof content === "object" && content && "type" in content && "props" in content;
+}
+
+const alignmentToJustify: Record<GridCellAlignment, Properties["justifyContent"]> = {
+  left: "flex-start",
+  center: "center",
+  right: "flex-end",
+};
+const alignmentToTextAlign: Record<GridCellAlignment, Properties["textAlign"]> = {
+  left: "left",
+  center: "center",
+  right: "right",
+};
+
+export function getAlignment(column: GridColumn<any>, maybeContent: ReactNode | GridCellContent): GridCellAlignment {
+  return (isGridCellContent(maybeContent) && maybeContent.alignment) || column.align || "left";
+}
+
+// For alignment, use: 1) cell def, else 2) column def, else 3) left.
+export function getJustification(
+  column: GridColumn<any>,
+  maybeContent: ReactNode | GridCellContent,
+  as: RenderAs,
+  alignment: GridCellAlignment,
+) {
+  // Always apply text alignment.
+  const textAlign = Css.add("textAlign", alignmentToTextAlign[alignment]).$;
+  if (as === "table") {
+    return textAlign;
+  }
+  return { ...Css.jc(alignmentToJustify[alignment]).$, ...textAlign };
+}
+
+/** Look at a row and get its filter value. */
+function filterValue(value: ReactNode | GridCellContent): any {
+  let maybeFn = value;
+  if (value && typeof value === "object") {
+    if ("value" in value) {
+      maybeFn = value.value;
+    } else if ("content" in value) {
+      maybeFn = value.content;
+    }
+  }
+  // Watch for functions that need to read from a potentially-changing proxy
+  if (maybeFn instanceof Function) {
+    return maybeFn();
+  }
+  return maybeFn;
+}
+
+export function maybeApplyFunction<T>(
+  row: T,
+  maybeFn: Properties | ((row: T) => Properties) | undefined,
+): Properties | undefined {
+  return typeof maybeFn === "function" ? maybeFn(row) : maybeFn;
+}
+
+export function matchesFilter(maybeContent: ReactNode | GridCellContent, filter: string): boolean {
+  let value = filterValue(maybeContent);
+  if (typeof value === "string") {
+    return value.toLowerCase().includes(filter.toLowerCase());
+  } else if (typeof value === "number") {
+    return Number(filter) === value;
+  }
+  return false;
+}
